+++ evosax/__init__.py
@@ -29,6 +29,7 @@ from .strategies import (
     GuidedES,
     ASEBO,
     CR_FM_NES,
+    MR15_GA,
 )
 from .utils import FitnessShaper, ParameterReshaper, ESLog
 from .networks import NetworkMapper
@@ -65,6 +66,7 @@ Strategies = {
     "GuidedES": GuidedES,
     "ASEBO": ASEBO,
     "CR_FM_NES": CR_FM_NES,
+    "MR15_GA": MR15_GA,
 }
 
 __all__ = [
@@ -106,4 +108,5 @@ __all__ = [
     "GuidedES",
     "ASEBO",
     "CR_FM_NES",
+    "MR15_GA",
 ]